Refactoring: created the useRatingModal hook
